requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML
openai==1.3.5
langchain==0.1.0
langchain-openai==0.0.2
transformers==4.35.2
torch==2.1.1
Pillow==10.1.0
sentence-transformers==2.2.2

# Vector Store
chromadb==0.4.18

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.26.2

# Streamlit (for frontend)
streamlit==1.50.2
requests==2.31.0